Move all basis functions to FE directory
