+++ kernel/Base/HpgemAPILinear_Impl.h
@@ -49,11 +49,11 @@
 #include "Integration/FaceIntegral.h"
 #include "Output/TecplotDiscontinuousSolutionWriter.h"
 #include "Output/TecplotSingleElementWriter.h"
-#include "Utilities/BasisFunctions1DH1ConformingLine.h"
-#include "Utilities/BasisFunctions2DH1ConformingSquare.h"
-#include "Utilities/BasisFunctions2DH1ConformingTriangle.h"
-#include "Utilities/BasisFunctions3DH1ConformingCube.h"
-#include "Utilities/BasisFunctions3DH1ConformingTetrahedron.h"
+#include "FE/BasisFunctions1DH1ConformingLine.h"
+#include "FE/BasisFunctions2DH1ConformingSquare.h"
+#include "FE/BasisFunctions2DH1ConformingTriangle.h"
+#include "FE/BasisFunctions3DH1ConformingCube.h"
+#include "FE/BasisFunctions3DH1ConformingTetrahedron.h"
 
 #include "Logger.h"
 namespace hpgem {
