+++ kernel/Base/MeshManipulator_Impl.h
@@ -71,17 +71,17 @@
 #include "L2Norm.h"
 #include "Geometry/Jacobian.h"
 #include "Geometry/ReferenceGeometry.h"
-#include "Utilities/BasisFunctions1DH1ConformingLine.h"
-#include "Utilities/BasisFunctions2DH1ConformingSquare.h"
-#include "Utilities/BasisFunctions2DH1ConformingTriangle.h"
-#include "Utilities/BasisFunctions2DNedelec.h"
-#include "Utilities/BasisFunctions3DH1ConformingCube.h"
-#include "Utilities/BasisFunctions3DH1ConformingPrism.h"
-#include "Utilities/BasisFunctions3DH1ConformingPyramid.h"
-#include "Utilities/BasisFunctions3DH1ConformingTetrahedron.h"
-#include "Utilities/BasisFunctions3DNedelec.h"
-#include "Utilities/BasisFunctions3DAinsworthCoyle.h"
-#include "Utilities/BasisFunctionsMonomials.h"
+#include "FE/BasisFunctions1DH1ConformingLine.h"
+#include "FE/BasisFunctions2DH1ConformingSquare.h"
+#include "FE/BasisFunctions2DH1ConformingTriangle.h"
+#include "FE/BasisFunctions2DNedelec.h"
+#include "FE/BasisFunctions3DH1ConformingCube.h"
+#include "FE/BasisFunctions3DH1ConformingPrism.h"
+#include "FE/BasisFunctions3DH1ConformingPyramid.h"
+#include "FE/BasisFunctions3DH1ConformingTetrahedron.h"
+#include "FE/BasisFunctions3DNedelec.h"
+#include "FE/BasisFunctions3DAinsworthCoyle.h"
+#include "FE/BasisFunctionsMonomials.h"
 #include "Logger.h"
 
 #include <algorithm>
