+++ kernel/FE/BasisFunctions1DH1ConformingLine.cpp
@@ -38,8 +38,8 @@
 
 #include "BasisFunctions1DH1ConformingLine.h"
 #include "helperFunctions.h"
-#include "FE/BasisFunctionSet.h"
-#include "FE/OrientedBasisFunctionSet.h"
+#include "BasisFunctionSet.h"
+#include "OrientedBasisFunctionSet.h"
 #include "Geometry/PointReference.h"
 
 // only uses the constant basis functions
