+++ kernel/FE/BasisFunctions2DNedelec.cpp
@@ -40,8 +40,8 @@
 
 #include <utility>
 
-#include "FE/BaseBasisFunction.h"
-#include "FE/BasisFunctionSet.h"
+#include "BaseBasisFunction.h"
+#include "BasisFunctionSet.h"
 #include "Geometry/PointReference.h"
 #include "Geometry/ReferenceTriangle.h"
