+++ kernel/FE/BasisFunctions3DH1ConformingPrism.cpp
@@ -38,8 +38,8 @@
 
 #include "BasisFunctions3DH1ConformingPrism.h"
 #include "helperFunctions.h"
-#include "FE/BasisFunctionSet.h"
-#include "FE/OrientedBasisFunctionSet.h"
+#include "BasisFunctionSet.h"
+#include "OrientedBasisFunctionSet.h"
 #include "Geometry/ReferenceTriangularPrism.h"
 #include "Geometry/PointReference.h"
