+++ kernel/FE/BasisFunctions3DNedelec.h
@@ -36,10 +36,10 @@
  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef HPGEM_KERNEL_BASISFUNCTIONS3DAINSWORTHCOYLE_H
-#define HPGEM_KERNEL_BASISFUNCTIONS3DAINSWORTHCOYLE_H
+#ifndef HPGEM_KERNEL_BASISFUNCTIONS3DNEDELEC_H
+#define HPGEM_KERNEL_BASISFUNCTIONS3DNEDELEC_H
 
-#include "FE/BaseBasisFunction.h"
+#include "BaseBasisFunction.h"
 #include <vector>
 
 namespace hpgem {
@@ -54,12 +54,13 @@ class PointReference;
 }
 
 namespace Utilities {
-//! Curl conforming edge functions.
-class BasisCurlEdgeAinsworthCoyle : public Base::BaseBasisFunction {
+
+//! Curl conforming Nedelec edge functions.
+class BasisCurlEdgeNedelec : public Base::BaseBasisFunction {
    public:
-    BasisCurlEdgeAinsworthCoyle(std::size_t degree,
-                                std::size_t localFirstVertex,
-                                std::size_t localSecondVertex);
+    BasisCurlEdgeNedelec(std::size_t degree1, std::size_t degree2,
+                         std::size_t localFirstVertex,
+                         std::size_t localSecondVertex);
 
     void eval(const Geometry::PointReference<3>& p,
               LinearAlgebra::SmallVector<3>& ret) const override;
@@ -68,15 +69,14 @@ class BasisCurlEdgeAinsworthCoyle : public Base::BaseBasisFunction {
         const Geometry::PointReference<3>& p) const override;
 
    private:
-    const std::size_t deg, o, i;
+    const std::size_t deg1, deg2, i, j;
 };
 
-//! Curl conforming edge based face functions.
-class BasisCurlEdgeFaceAinsworthCoyle : public Base::BaseBasisFunction {
+//! Curl conforming Nedelec face functions.
+class BasisCurlFace1Nedelec : public Base::BaseBasisFunction {
    public:
-    BasisCurlEdgeFaceAinsworthCoyle(std::size_t degree,
-                                    std::size_t localOpposingVertex,
-                                    std::size_t localSpecialVertex);
+    BasisCurlFace1Nedelec(std::size_t degree1, std::size_t degree2,
+                          std::size_t degree3, std::size_t localOpposingVertex);
 
     void eval(const Geometry::PointReference<3>& p,
               LinearAlgebra::SmallVector<3>& ret) const override;
@@ -85,15 +85,13 @@ class BasisCurlEdgeFaceAinsworthCoyle : public Base::BaseBasisFunction {
         const Geometry::PointReference<3>& p) const override;
 
    private:
-    std::size_t deg, a, b, c;
+    std::size_t deg1, deg2, deg3, a, b, c, d;
 };
 
-//! Curl conforming face functions.
-class BasisCurlFaceAinsworthCoyle : public Base::BaseBasisFunction {
+class BasisCurlFace2Nedelec : public Base::BaseBasisFunction {
    public:
-    BasisCurlFaceAinsworthCoyle(std::size_t degree1, std::size_t degree2,
-                                std::size_t localOpposingVertex,
-                                std::size_t direction);
+    BasisCurlFace2Nedelec(std::size_t degree1, std::size_t degree2,
+                          std::size_t degree3, std::size_t localOpposingVertex);
 
     void eval(const Geometry::PointReference<3>& p,
               LinearAlgebra::SmallVector<3>& ret) const override;
@@ -102,15 +100,14 @@ class BasisCurlFaceAinsworthCoyle : public Base::BaseBasisFunction {
         const Geometry::PointReference<3>& p) const override;
 
    private:
-    std::size_t deg1, deg2, a, b, c;
+    std::size_t deg1, deg2, deg3, a, b, c, d;
 };
 
-//! Curl conforming face based interior functions.
-class BasisCurlFaceinteriorAinsworthCoyle : public Base::BaseBasisFunction {
+//! curl conforming interior Nedelec functions
+class BasisCurlinterior1Nedelec : public Base::BaseBasisFunction {
    public:
-    BasisCurlFaceinteriorAinsworthCoyle(std::size_t degree1,
-                                        std::size_t degree2,
-                                        std::size_t localOpposingVertex);
+    BasisCurlinterior1Nedelec(std::size_t degree1, std::size_t degree2,
+                              std::size_t degree3, std::size_t degree4);
 
     void eval(const Geometry::PointReference<3>& p,
               LinearAlgebra::SmallVector<3>& ret) const override;
@@ -119,14 +116,14 @@ class BasisCurlFaceinteriorAinsworthCoyle : public Base::BaseBasisFunction {
         const Geometry::PointReference<3>& p) const override;
 
    private:
-    std::size_t deg1, deg2, a, b, c, d;
+    const std::size_t deg1, deg2, deg3, deg4;
 };
 
-//! curl conforming interior functions
-class BasisCurlinteriorAinsworthCoyle : public Base::BaseBasisFunction {
+//! curl conforming interior Nedelec functions
+class BasisCurlinterior2Nedelec : public Base::BaseBasisFunction {
    public:
-    BasisCurlinteriorAinsworthCoyle(std::size_t degree1, std::size_t degree2,
-                                    std::size_t degree3, std::size_t direction);
+    BasisCurlinterior2Nedelec(std::size_t degree1, std::size_t degree2,
+                              std::size_t degree3, std::size_t degree4);
 
     void eval(const Geometry::PointReference<3>& p,
               LinearAlgebra::SmallVector<3>& ret) const override;
@@ -135,14 +132,29 @@ class BasisCurlinteriorAinsworthCoyle : public Base::BaseBasisFunction {
         const Geometry::PointReference<3>& p) const override;
 
    private:
-    const std::size_t deg1, deg2, deg3, direction;
+    const std::size_t deg1, deg2, deg3, deg4;
 };
 
-Base::BasisFunctionSet* createDGBasisFunctionSet3DAinsworthCoyle(
-    std::size_t order);
+//! curl conforming interior Nedelec functions
+class BasisCurlinterior3Nedelec : public Base::BaseBasisFunction {
+   public:
+    BasisCurlinterior3Nedelec(std::size_t degree1, std::size_t degree2,
+                              std::size_t degree3, std::size_t degree4);
+
+    void eval(const Geometry::PointReference<3>& p,
+              LinearAlgebra::SmallVector<3>& ret) const override;
+
+    LinearAlgebra::SmallVector<3> evalCurl(
+        const Geometry::PointReference<3>& p) const override;
+
+   private:
+    const std::size_t deg1, deg2, deg3, deg4;
+};
+
+Base::BasisFunctionSet* createDGBasisFunctionSet3DNedelec(std::size_t order);
 
 }  // namespace Utilities
 
 }  // namespace hpgem
 
-#endif  // HPGEM_KERNEL_BASISFUNCTIONS3DAINSWORTHCOYLE_H
+#endif  // HPGEM_KERNEL_BASISFUNCTIONS3DNEDELEC_H
