+++ kernel/FE/BasisFunctionsMonomials.cpp
@@ -37,8 +37,8 @@
 */
 
 #include "BasisFunctionsMonomials.h"
-#include "FE/BaseBasisFunction.h"
-#include "FE/BasisFunctionSet.h"
+#include "BaseBasisFunction.h"
+#include "BasisFunctionSet.h"
 #include "Geometry/PointReference.h"
 
 #include <array>
