+++ kernel/FE/helperFunctions.cpp
