+++ kernel/FE/helperFunctions.h
