+++ tests/self/Basic/060MultipleVariables_SelfTest.cpp
@@ -47,7 +47,7 @@
 
 #include <iostream>
 
-#include "Utilities/BasisFunctionsMonomials.h"
+#include "FE/BasisFunctionsMonomials.h"
 #include "Base/Element.h"
 #include "Base/Face.h"
 #include "Base/Node.h"
